Move Cursor interface location
